refactor(UserAccountScreen): Don't save locations with blank addresses.
