+++ lib/components/user/favorite-locations-pane.js
@@ -1,4 +1,3 @@
-import clone from 'clone'
 import { Field, FieldArray } from 'formik'
 import memoize from 'lodash.memoize'
 import React, { Component } from 'react'
@@ -38,39 +37,8 @@ const NewLocationFormControl = styled(FormControl)`
 `
 
 // Helper filter functions.
-const isHome = loc => loc.type === 'home'
-const isWork = loc => loc.type === 'work'
-const notHomeOrWork = loc => loc.type !== 'home' && loc.type !== 'work'
-
-/**
- * Makes a copy of the logged-in user data for the Formik initial state,
- * with the 'home' and 'work' locations at the top of the savedLocations list
- * so they are always shown and shown at the top of the FavoriteLocationsPane.
- * Note: In the returned value, savedLocations is always a valid array.
- */
-export function cloneWithHomeAndWorkAsTopLocations (loggedInUser) {
-  const clonedUser = clone(loggedInUser)
-  const { savedLocations = [] } = clonedUser
-
-  const homeLocation = savedLocations.find(isHome) || {
-    address: '',
-    icon: 'home',
-    type: 'home'
-  }
-  const workLocation = savedLocations.find(isWork) || {
-    address: '',
-    icon: 'briefcase',
-    type: 'work'
-  }
-  const reorderedLocations = [
-    homeLocation,
-    workLocation,
-    ...savedLocations.filter(notHomeOrWork)
-  ]
-
-  clonedUser.savedLocations = reorderedLocations
-  return clonedUser
-}
+export const isHome = loc => loc.type === 'home'
+export const isWork = loc => loc.type === 'work'
 
 function addNewAddress (arrayHelpers, e) {
   const value = (e.target.value || '').trim()
