+++ lib/components/user/user-account-screen.js
@@ -1,3 +1,4 @@
+import clone from 'clone'
 import { Form, Formik } from 'formik'
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
@@ -10,7 +11,7 @@ import { isNewUser } from '../../util/user'
 import DesktopNav from '../app/desktop-nav'
 import AccountSetupFinishPane from './account-setup-finish-pane'
 import ExistingAccountDisplay from './existing-account-display'
-import FavoriteLocationsPane, { cloneWithHomeAndWorkAsTopLocations } from './favorite-locations-pane'
+import FavoriteLocationsPane, { isHome, isWork } from './favorite-locations-pane'
 import NewAccountWizard from './new-account-wizard'
 import NotificationPrefsPane from './notification-prefs-pane'
 import PhoneVerificationPane from './phone-verification-pane'
@@ -33,18 +34,46 @@ const validationSchema = yup.object({
 })
 
 /**
+ * Makes a copy of the logged-in user data for the Formik initial state,
+ * with the 'home' and 'work' locations at the top of the savedLocations list
+ * so they are always shown and shown at the top of the FavoriteLocationsPane.
+ * Note: In the returned value, savedLocations is always a valid array.
+ */
+function cloneWithHomeAndWorkAsTopLocations (loggedInUser) {
+  const clonedUser = clone(loggedInUser)
+  const { savedLocations = [] } = clonedUser
+
+  const homeLocation = savedLocations.find(isHome) || {
+    address: '',
+    icon: 'home',
+    type: 'home'
+  }
+  const workLocation = savedLocations.find(isWork) || {
+    address: '',
+    icon: 'briefcase',
+    type: 'work'
+  }
+  const reorderedLocations = [
+    homeLocation,
+    workLocation,
+    ...savedLocations.filter(loc => loc !== homeLocation && loc !== workLocation)
+  ]
+
+  clonedUser.savedLocations = reorderedLocations
+  return clonedUser
+}
+
+/**
  * This screen handles creating/updating OTP user account settings.
  */
 class UserAccountScreen extends Component {
   _updateUserPrefs = async userData => {
     // TODO: Change state of Save button while the update action takes place.
 
-    // FIXME: Filter out blank 'home' and 'work' locations?
-    // (That used to be done in the favorite locations pane.)
-    // If not, they will be saved, however there is no side effect.
-
-    const { createOrUpdateUser } = this.props
-    await createOrUpdateUser(userData)
+    // In userData.savedLocations, filter out entries with blank addresses.
+    const newUserData = clone(userData)
+    newUserData.savedLocations = newUserData.savedLocations.filter(({ address }) => address && address.length)
+    await this.props.createOrUpdateUser(newUserData)
 
     // TODO: Handle UI feedback (currently an alert() dialog inside createOrUpdateUser).
   }
